requirements.txt
hydra-core>=1.1.1
torch==1.13.1
torchvision==0.14.1
tqdm==4.64.1
typing==3.7.4.3
Pillow==9.3.0
PyYAML==6.0
argparse==1.4.0
boto3==1.34.51
python-dotenv==1.0.0